# docs/site/content/en/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Horreum REST API
  description: "Horreum automated change anomaly detection. For more information,\
    \ please see [https://horreum.hyperfoil.io/](https://horreum.hyperfoil.io/)"
  version: "0.13"
...
paths:
...
  /api/schema:
    get:
      tags:
      - Schema
      description: Retrieve a paginated list of Schemas with available count
      operationId: listSchemas
      parameters:
      - name: limit
        in: query
        description: limit the number of results
        schema:
          format: int32
          type: integer
        example: 20
      - name: page
        in: query
        description: filter by page number of a paginated list of Schemas
        schema:
          format: int32
          type: integer
        example: 2
      - name: sort
        in: query
        description: Field name to sort results
        schema:
          type: string
        example: name
      - name: direction
        in: query
        description: Sort direction
        schema:
          $ref: '#/components/schemas/SortDirection'
        example: Ascending
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaQueryResult'
    post:
      tags:
      - Schema
      description: Save a new Schema
      operationId: addSchema
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schema'
      responses:
        "200":
          description: Import a new Schema
          content:
            application/json:
              schema:
                format: int32
                type: integer
              example: 103
  /api/schema/allLabels:
    get:
      tags:
      - Schema
      description: Retrieve list of Labels for ny name. Allows users to retrieve all
        Label Definitions that have the same name
      operationId: allLabels
      parameters:
      - name: name
        in: query
        description: Label name
        schema:
          type: string
        example: buildID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LabelInfo'
...
  /api/schema/descriptors:
    get:
      tags:
      - Schema
      description: Retrieve a list of Schema Descriptors
      operationId: descriptors
      parameters:
      - name: id
        in: query
        description: Limit to a single Schema by ID
        schema:
          type: array
          items:
            format: int32
            type: integer
        example: 102
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchemaDescriptor'
  /api/schema/findUsages:
    get:
      tags:
      - Schema
      description: Find all usages of a Schema by label name
      operationId: findUsages
      parameters:
      - name: label
        in: query
        description: Name of label to search for
        required: true
        schema:
          type: string
        example: Throughput
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LabelLocation'
  /api/schema/idByUri/{uri}:
    get:
      tags:
      - Schema
      description: Retrieve Schema ID by uri
      operationId: idByUri
      parameters:
      - name: uri
        in: path
        description: Schema uri
        required: true
        schema:
          type: string
        example: uri:my-schema:0.1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                format: int32
                type: integer
              example: 101
  /api/schema/import:
    post:
      tags:
      - Schema
      description: Import an previously exported Schema either as a new Schema or
        to update an existing Schema
      operationId: importSchema
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaExport'
        required: true
      responses:
        "201":
          description: Import a new Schema or update an existing Schema
  /api/schema/{id}:
    get:
      tags:
      - Schema
      description: Retrieve Schema by ID
      operationId: getSchema
      parameters:
      - name: id
        in: path
        description: Schema ID to retrieve
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      - name: token
        in: query
        description: API token for authorization
        schema:
          type: string
        example: 101
      responses:
        "404":
          description: No Schema with the given id was found
          content:
            application/json: {}
        "200":
          description: Returns Schema if a matching id is found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schema'
    delete:
      tags:
      - Schema
      description: Delete a Schema by id
      operationId: deleteSchema
      parameters:
      - name: id
        in: path
        description: Schema ID to delete
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      responses:
        "204":
          description: No Content
  /api/schema/{id}/dropToken:
    delete:
      tags:
      - Schema
      description: Remove access token for schema
      operationId: dropSchemaToken
      parameters:
      - name: id
        in: path
        description: Token ID
        required: true
        schema:
          format: int32
          type: integer
        example: 102
      responses:
        "204":
          description: No Content
  /api/schema/{id}/export:
    get:
      tags:
      - Schema
      description: Export a Schema
      operationId: exportSchema
      parameters:
      - name: id
        in: path
        description: Schema ID
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      responses:
        "200":
          description: A JSON representation of the SchemaExport object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaExport'
  /api/schema/{id}/resetToken:
    post:
      tags:
      - Schema
      description: Regenerate access token for schema
      operationId: resetSchemaToken
      parameters:
      - name: id
        in: path
        description: Token ID
        required: true
        schema:
          format: int32
          type: integer
        example: 102
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
              example: 094678029a2aaf9a2847502273099bb3a1b2338c2b9c618ed09aef0181666e38
  /api/schema/{id}/updateAccess:
    post:
      tags:
      - Schema
      description: Update the Access configuration for a Schema
      operationId: updateSchemaAccess
      parameters:
      - name: id
        in: path
        description: Schema ID to update Access
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      - name: owner
        in: query
        description: Name of the new owner
        required: true
        schema:
          type: string
        example: perf-team
      - name: access
        in: query
        description: New Access level
        required: true
        schema:
          $ref: '#/components/schemas/Access'
        example: 0
      responses:
        "201":
          description: Created
  /api/schema/{schemaId}/labels:
    get:
      tags:
      - Schema
      description: Retrieve list of Labels for a Schema by Schema ID
      operationId: labels
      parameters:
      - name: schemaId
        in: path
        description: Schema ID
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Label'
    post:
      tags:
      - Schema
      description: Save new or update existing Label for a Schema
      operationId: addOrUpdateLabel
      parameters:
      - name: schemaId
        in: path
        description: Schema ID
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Label'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                format: int32
                type: integer
  /api/schema/{schemaId}/labels/{labelId}:
    delete:
      tags:
      - Schema
      description: Delete existing Label from a Schema
      operationId: deleteLabel
      parameters:
      - name: schemaId
        in: path
        description: Schema ID
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      - name: labelId
        in: path
        description: Label ID
        required: true
        schema:
          format: int32
          type: integer
        example: 202
      responses:
        "204":
          description: No Content
  /api/schema/{schemaId}/transformers:
    get:
      tags:
      - Schema
      description: List all Transformers defined for a Schema
      operationId: listTransformers
      parameters:
      - name: schemaId
        in: path
        description: Schema ID
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transformer'
    post:
      tags:
      - Schema
      description: Save new or update existing Transformer defintion
      operationId: addOrUpdateTransformer
      parameters:
      - name: schemaId
        in: path
        description: Schema ID
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transformer'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                format: int32
                type: integer
  /api/schema/{schemaId}/transformers/{transformerId}:
    delete:
      tags:
      - Schema
      description: Delete a Transformer defined for a Schema
      operationId: deleteTransformer
      parameters:
      - name: schemaId
        in: path
        description: Schema ID
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      - name: transformerId
        in: path
        description: Transformer ID
        required: true
        schema:
          format: int32
          type: integer
        example: 202
      responses:
        "204":
          description: No Content
...
components:
  schemas:
    Access:
      description: "Resources have different visibility within the UI. 'PUBLIC', 'PROTECTED'\
        \ and 'PRIVATE'. Restricted resources are not visible to users who do not\
        \ have the correct permissions"
      enum:
      - PUBLIC
      - PROTECTED
      - PRIVATE
      type: string
...
    LabelInfo:
      required:
      - name
      - metrics
      - filtering
      - schemas
      type: object
      properties:
        name:
          description: Label name
          type: string
          example: buildID
        metrics:
          description: Is label a metrics label?
          type: boolean
          example: true
        filtering:
          description: Is label a filtering label?
          type: boolean
          example: false
        schemas:
          description: List of schemas where label is referenced
          type: array
          items:
            $ref: '#/components/schemas/SchemaDescriptor'
    LabelLocation:
      type: object
      properties:
        type:
          description: Location of Label usage
          type: string
          example: VIEW
        testId:
          format: int32
          description: Unique ID for location that references Schema
          type: integer
          example: 101
        testName:
          description: Test name that references Schema
          type: string
          example: My Benchmark
...
    Schema:
      description: Data object that describes the schema definition for a test
      required:
      - id
      - uri
      - name
      type: object
      allOf:
      - $ref: '#/components/schemas/ProtectedType'
      properties:
        id:
          format: int32
          description: Unique Schema ID
          type: integer
          example: 101
        uri:
          description: "Unique, versioned schema URI"
          type: string
          example: uri:my-schema:0.1
        name:
          description: Schema name
          type: string
          example: My benchmark schema
        description:
          description: Schema Description
          type: string
          example: Schema for processing my benchmark
        schema:
          description: JSON validation schema. Used to validate uploaded JSON documents
          type: string
          example: "{  \"$schema\": \"https://json-schema.org/draft/2020-12/schema\"\
            ,  \"$id\": \"https://example.com/product.schema.json\",  \"title\": \"\
            Product\",  \"description\": \"A product in the catalog\",  \"type\":\
            \ \"object\"}"
        token:
          description: Array of API tokens associated with test
          type: string
          example: ""
    SchemaDescriptor:
      required:
      - id
      - name
      - uri
      type: object
      properties:
        id:
          format: int32
          description: Schema unique ID
          type: integer
          example: 1
        name:
          description: Schema name
          type: string
          example: my-benchmark-schema
        uri:
          description: Schema name
          type: string
          example: uri:my-schmea:0.1
    SchemaExport:
      description: Represents a Schema with all associated data used for export/import
        operations.
      type: object
      allOf:
      - $ref: '#/components/schemas/Schema'
      properties:
        labels:
          description: Array of Labels associated with schema
          type: array
          items:
            $ref: '#/components/schemas/Label'
        transformers:
          description: Array of Transformers associated with schema
          type: array
          items:
            $ref: '#/components/schemas/Transformer'
    SchemaQueryResult:
      required:
      - schemas
      - count
      type: object
      properties:
        schemas:
          description: Array of Schemas
          type: array
          items:
            $ref: '#/components/schemas/Schema'
        count:
          format: int64
          description: Count of available Schemas. This is a count of Schemas that
            the current user has access to
          type: integer
          example: 64
...
    SortDirection:
      enum:
      - Ascending
      - Descending
      type: string
...